Reorganized the CommandOperationHelper
Added SyncCommandOperationHelper and AsyncCommandOperationHelper.
Improved readibility of the classes. Also makes refactoring methods
in the future simplier.

Finally, the refactoring identified methods that weren't used in
the operations at all or duplicates. These have now been removed.
