+++ driver-core/src/main/com/mongodb/internal/operation/CreateIndexesOperation.java
@@ -50,10 +50,10 @@ import static com.mongodb.internal.async.ErrorHandlingResultCallback.errorHandli
 import static com.mongodb.internal.operation.AsyncOperationHelper.AsyncCallableWithConnection;
 import static com.mongodb.internal.operation.AsyncOperationHelper.releasingCallback;
 import static com.mongodb.internal.operation.AsyncOperationHelper.withAsyncConnection;
-import static com.mongodb.internal.operation.CommandOperationHelper.executeCommand;
-import static com.mongodb.internal.operation.CommandOperationHelper.executeCommandAsync;
-import static com.mongodb.internal.operation.CommandOperationHelper.writeConcernErrorTransformer;
-import static com.mongodb.internal.operation.CommandOperationHelper.writeConcernErrorWriteTransformer;
+import static com.mongodb.internal.operation.SyncCommandOperationHelper.executeCommand;
+import static com.mongodb.internal.operation.AsyncCommandOperationHelper.executeCommandAsync;
+import static com.mongodb.internal.operation.SyncCommandOperationHelper.writeConcernErrorTransformer;
+import static com.mongodb.internal.operation.AsyncCommandOperationHelper.writeConcernErrorTransformerAsync;
 import static com.mongodb.internal.operation.DocumentHelper.putIfNotZero;
 import static com.mongodb.internal.operation.IndexHelper.generateIndexName;
 import static com.mongodb.internal.operation.OperationHelper.LOGGER;
@@ -221,7 +221,7 @@ public class CreateIndexesOperation implements AsyncWriteOperation<Void>, WriteO
                             } else {
                                 try {
                                     executeCommandAsync(binding, namespace.getDatabaseName(),
-                                            getCommand(connection.getDescription()), connection, writeConcernErrorWriteTransformer(),
+                                            getCommand(connection.getDescription()), connection, writeConcernErrorTransformerAsync(),
                                             new SingleResultCallback<Void>() {
                                                 @Override
                                                 public void onResult(final Void result, final Throwable t) {
