+++ driver-core/src/main/com/mongodb/internal/operation/CreateViewOperation.java
@@ -35,10 +35,10 @@ import java.util.List;
 
 import static com.mongodb.assertions.Assertions.notNull;
 import static com.mongodb.internal.async.ErrorHandlingResultCallback.errorHandlingCallback;
-import static com.mongodb.internal.operation.CommandOperationHelper.executeCommand;
-import static com.mongodb.internal.operation.CommandOperationHelper.executeCommandAsync;
-import static com.mongodb.internal.operation.CommandOperationHelper.writeConcernErrorTransformer;
-import static com.mongodb.internal.operation.CommandOperationHelper.writeConcernErrorWriteTransformer;
+import static com.mongodb.internal.operation.SyncCommandOperationHelper.executeCommand;
+import static com.mongodb.internal.operation.AsyncCommandOperationHelper.executeCommandAsync;
+import static com.mongodb.internal.operation.SyncCommandOperationHelper.writeConcernErrorTransformer;
+import static com.mongodb.internal.operation.AsyncCommandOperationHelper.writeConcernErrorTransformerAsync;
 import static com.mongodb.internal.operation.OperationHelper.LOGGER;
 import static com.mongodb.internal.operation.AsyncOperationHelper.releasingCallback;
 import static com.mongodb.internal.operation.AsyncOperationHelper.withAsyncConnection;
@@ -173,7 +173,7 @@ public class CreateViewOperation implements AsyncWriteOperation<Void>, WriteOper
                         wrappedCallback.onResult(null, createExceptionForIncompatibleServerVersion());
                     }
                     executeCommandAsync(binding, databaseName, getCommand(connection.getDescription()),
-                            connection, writeConcernErrorWriteTransformer(), wrappedCallback);
+                            connection, writeConcernErrorTransformerAsync(), wrappedCallback);
                 }
             }
         });
