+++ driver-core/src/main/com/mongodb/internal/operation/FindAndModifyHelper.java
@@ -21,8 +21,8 @@ import com.mongodb.ServerAddress;
 import com.mongodb.WriteConcernResult;
 import com.mongodb.internal.connection.AsyncConnection;
 import com.mongodb.internal.connection.Connection;
-import com.mongodb.internal.operation.CommandOperationHelper.CommandWriteTransformer;
-import com.mongodb.internal.operation.CommandOperationHelper.CommandWriteTransformerAsync;
+import com.mongodb.internal.operation.SyncCommandOperationHelper.CommandWriteTransformer;
+import com.mongodb.internal.operation.AsyncCommandOperationHelper.CommandWriteTransformerAsync;
 import org.bson.BsonBoolean;
 import org.bson.BsonDocument;
 import org.bson.BsonInt32;
